network-transport T2: the letter owns its updates by value, and the leak ends
m_updates was a std::vector<NetworkUpdate*> of raw owning pointers in a class
with no destructor, so every Update letter the client received and every
history copy the server made leaked all of its updates. Counted in the g++
harness with a global operator new/delete pair: three letters carrying eight
updates each — one built, one broadcast copy, one parsed off the wire — leak
24 allocations before this change and zero after.

Holding them by value deletes the leak, the hand-written copy constructor,
the new-plus-memcpy in AddUpdate, and the "remember to update the copy
constructor" warning in one move. NetworkUpdate's own operator= goes too: it
was memcpy(this, &n, sizeof(*this)), which is what the implicit one does for
a class of scalars, and declaring it was the only thing stopping the type
being trivially copyable.

The letter's GetByteStream becomes Serialise(char*, int) into a caller-owned
buffer. It used to write into one file-scope char[1024] shared by every
letter in the process — safe only because the single caller sent immediately,
and about to stop being safe at all once T4 and T5 change who serialises.

Bytes on the wire are unchanged; the harness still compares every update and
letter type against the pre-T1 macro serialisers byte for byte.

Not compiled by MSVC and not run as a suite. See the task's notes.

+++ Tests/NeuronCoreTests/NetworkUpdateTests.cpp
@@ -421,7 +421,7 @@ namespace NeuronCoreTests
         writer.Write<int>(31337);
         writer.Write<int>(0);
 
-        const NetworkUpdate received(datagram, sizeof(datagram));
+        const NetworkUpdate received(datagram, static_cast<int>(sizeof(datagram)));
 
         Assert::IsFalse(received.IsValid());
       }
